add CUDA extension
